Pull out module processor function into util file
