+++ tasks/postcss.ts
@@ -1,38 +1,12 @@
-const path = require('path');
-const fs = require('fs');
-const postCssImport = require('postcss-import');
-const postCssNext = require('postcss-cssnext');
-const postCssModules = require('postcss-modules');
-const umdWrapper = require('./util/umdWrapper');
+import { createProcessors } from './util/postcss';
 
-function moduleProcessors(dest: string, cwd = '', dist?: boolean) {
-	const scopedName = dist ? '[hash:base64:8]' : '[name]__[local]__[hash:base64:5]' ;
-	return [
-		postCssImport,
-		postCssNext({
-			features: {
-				autoprefixer: {
-					browsers: [
-						'last 2 versions',
-						'ie >= 10'
-					]
-				}
-			}
-		}),
-		postCssModules({
-			generateScopedName: scopedName,
-			getJSON: function(cssFileName: string, json: any) {
-				const outputPath = path.resolve(dest, path.relative(cwd, cssFileName));
-				const newFilePath = outputPath + '.js';
-				const themeKey = ' _key';
-				json[themeKey] = 'dojo-' + path.basename(outputPath, '.css');
-				fs.writeFileSync(newFilePath, umdWrapper(JSON.stringify(json)));
-			}
-		})
-	];
-};
-
-function init(grunt: IGrunt) {
+export = function init(grunt: IGrunt) {
+	const path = require('path');
+	const fs = require('fs');
+	const postCssImport = require('postcss-import');
+	const postCssNext = require('postcss-cssnext');
+	const postCssModules = require('postcss-modules');
+	const umdWrapper = require('./util/umdWrapper');
 	grunt.loadNpmTasks('grunt-postcss');
 
 	const distDirectory = grunt.config.get<string>('distDirectory') || '';
@@ -72,13 +46,13 @@ function init(grunt: IGrunt) {
 		'modules-dev': {
 			files: moduleFiles(path.join(devDirectory, 'src')),
 			options: {
-				processors: moduleProcessors(devDirectory)
+				processors: createProcessors(devDirectory)
 			}
 		},
 		'modules-dist': {
 			files: moduleFiles(distDirectory),
 			options: {
-				processors: moduleProcessors(distDirectory, 'src', true)
+				processors: createProcessors(distDirectory, 'src', true)
 			}
 		},
 		variables: {
@@ -88,9 +62,4 @@ function init(grunt: IGrunt) {
 			}
 		}
 	});
-};
-
-module.exports = {
-	moduleProcessors,
-	init
 };
